fixed eq bug, kick out restack, cleanup example
